pyproject.toml
[project]
name = "revou-bank"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    "bcrypt>=4.3.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2>=2.9.10",
    "pydantic>=2.10.6",
    "taskipy>=1.14.1",
]
[tool.taskipy.tasks]
fr = "flask --app main run --port 8000 --reload --debug"
